Update 2019-03-10-abou-moustafa19a.md
The paper title in the PDF file is the correct title: "An Exponential Efron-Stein Inequality for Lq Stable Learning Rules". 
The current title on the PMLR website is showing a different one: "An Exponential Tail bound for Lq Stable Learning Rules".  
We will appreciate it if you can change the PMLR website title to be consistent with the actual title for the paper.